Fix MSVC compiler warnings across codebase
- Fix C4244 warnings (type conversion/data loss) by adding explicit casts
  for u32->u16, int->i8 conversions in aura_processor, registry_combat_resolver,
  ai_controller, game_runner, result_analyzer, and batch_simulator headers

- Fix C4100 warnings (unreferenced parameters) by commenting out unused
  parameters in registry_combat_resolver, spell_caster, text_match_logger,
  export_game_runner, chunk_sim, main.cpp, and combat_rules.cpp

- Fix C4189 warnings (unused local variables) by removing unused variables
  in ai_controller, rule_aware_ai, spell_caster, batch_simulator, batch_sim,
  chunk_sim, and opr_pipeline

- Fix C4996 warnings (deprecated functions) by replacing:
  - gmtime with gmtime_s/gmtime_r in game_event_log.hpp
  - strcpy with FixedString constructor in test files

- Fix ::tolower warnings by using proper lambda with unsigned char cast
  in std::transform calls across unit_parser, opr_pipeline.hpp and .cpp

- Add default cases to switch statements to handle all enum values in
  batch_simulator.hpp

+++ include/core/aura_processor.hpp
@@ -88,7 +88,7 @@ public:
                         ActiveModifier mod(aura.grants_rule, aura.grants_value);
                         mod.source = ModifierSource::AURA;
                         mod.duration = ModifierDuration::UNTIL_END_OF_PHASE;
-                        mod.source_unit_id = unit_a.unit_id;
+                        mod.source_unit_id = static_cast<u16>(unit_a.unit_id);
                         mod.condition = aura.condition;
                         mods_b.add_modifier(mod);
                         ++applied;
@@ -100,7 +100,7 @@ public:
                     ActiveModifier mod(aura.grants_rule, aura.grants_value);
                     mod.source = ModifierSource::AURA;
                     mod.duration = ModifierDuration::UNTIL_END_OF_PHASE;
-                    mod.source_unit_id = unit_a.unit_id;
+                    mod.source_unit_id = static_cast<u16>(unit_a.unit_id);
                     mod.condition = aura.condition;
                     mods_a.add_modifier(mod);
                     ++applied;
@@ -124,7 +124,7 @@ public:
                         ActiveModifier mod(aura.grants_rule, aura.grants_value);
                         mod.source = ModifierSource::AURA;
                         mod.duration = ModifierDuration::UNTIL_END_OF_PHASE;
-                        mod.source_unit_id = unit_b.unit_id;
+                        mod.source_unit_id = static_cast<u16>(unit_b.unit_id);
                         mod.condition = aura.condition;
                         mods_a.add_modifier(mod);
                         ++applied;
@@ -136,7 +136,7 @@ public:
                     ActiveModifier mod(aura.grants_rule, aura.grants_value);
                     mod.source = ModifierSource::AURA;
                     mod.duration = ModifierDuration::UNTIL_END_OF_PHASE;
-                    mod.source_unit_id = unit_b.unit_id;
+                    mod.source_unit_id = static_cast<u16>(unit_b.unit_id);
                     mod.condition = aura.condition;
                     mods_b.add_modifier(mod);
                     ++applied;
